Split tests into own module
